backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-groq
python-dotenv
faiss-cpu
pydantic
pypdf
sentence-transformers
groq